scanner/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hipaa-scanner"
version = "1.0.0"
description = "HIPAA compliance scanner for Windows networks — MSP edition"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "impacket>=0.11",
    "ldap3>=2.9",
    "pywinrm>=0.4",
    "python-nmap>=0.7",
    "keyring>=24.0",
    "cryptography>=42.0",
    "pydantic>=2.5",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
hipaa-scan = "hipaa_scanner.cli:main"

[tool.setuptools.packages.find]
where = ["."]
